pyproject.toml
[project]
name = "rag-nano"
version = "0.1.0"
description = "RAG Nano v1 — minimum closed loop librarian: ingest + retrieval + evaluation, agent-neutral HTTP API."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "rag-nano" }]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sentence-transformers>=2.7",
    "numpy>=1.26",
    "typer>=0.12",
    "pyyaml>=6.0",
    "python-ulid>=2.7",
]

[project.scripts]
rag-nano = "rag_nano.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_nano"]

[dependency-groups]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "httpx>=0.27",
    "ruff>=0.4",
    "pyright>=1.1.360",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["tests/fixtures"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # simplify
]
ignore = [
    "E501",  # line length handled by formatter
    "UP042", # StrEnum migration is behavior-changing for serialization; defer
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B"]
# FastAPI Depends() / Typer Argument() / Typer Option() in argument defaults is the canonical idiom.
"rag_nano/api/routes.py" = ["B008"]
"rag_nano/cli/main.py" = ["B008"]

[tool.pyright]
include = ["rag_nano"]
exclude = ["**/__pycache__", "**/.venv", "tests/fixtures"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = "error"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-q",
    "--strict-markers",
    "--strict-config",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*",
]
